added circuit support
